Cab Voucher controller modified, Dashboard controller modified, view files modified

+++ database/migrations/2023_11_01_095544_create_categories_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
             $table->string('categoryName')->nullable();
-            $table->integer('isActive')->default(1);
+            $table->tinyInteger('isActive')->default(1);
         });
     }
